LayoutTests/media/track/track-webvtt-tc004-magic-header-expected.txt

Tests that the magic file header "WEBVTT" leads to the file properly recognized as a WebVTT file.


*** Testing text track 0
EXPECTED (cues.length == '2') OK
EXPECTED (cues[i].id == '1') OK
EXPECTED (cues[i].startTime == '0') OK
EXPECTED (cues[i].endTime == '30.5') OK
EXPECTED (cues[i].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (cues[i].id == '2') OK
EXPECTED (cues[i].startTime == '31') OK
EXPECTED (cues[i].endTime == '1200.5') OK
EXPECTED (cues[i].getCueAsSource() == 'I said Bear is coming!!!!') OK

*** Testing text track 1
EXPECTED (cues.length == '2') OK
EXPECTED (cues[i].id == '1') OK
EXPECTED (cues[i].startTime == '0') OK
EXPECTED (cues[i].endTime == '30.5') OK
EXPECTED (cues[i].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (cues[i].id == '2') OK
EXPECTED (cues[i].startTime == '31') OK
EXPECTED (cues[i].endTime == '1200.5') OK
EXPECTED (cues[i].getCueAsSource() == 'I said Bear is coming!!!!') OK

*** Testing text track 2
EXPECTED (cues.length == '0') OK
END OF TEST
